requirements.txt
streamlit
anthropic
python-dotenv
elevenlabs
requests
chromadb
sentence-transformers
audio-recorder-streamlit
numpy